services/llm_service/providers/openai_provider.py
"""
OpenAI LLM Provider Implementation (TEMPLATE)

Date: 2025-12-11
Author: BabySteps Development Team

Purpose:
    Template implementation of LLMProvider for OpenAI.
    Shows how to add support for GPT-4, GPT-3.5-turbo, etc.

Status:
    TEMPLATE - Uncomment and configure when you want to use OpenAI

Setup:
    1. Install: pip install openai
    2. Set API key in environment: OPENAI_API_KEY=your-key
    3. Update settings.py: LLM_PROVIDER = 'openai'
    4. Uncomment this file's import in providers/__init__.py

Usage:
    llm = get_llm_provider()  # Will use OpenAI if configured
    response = llm.chat(message="Hello", system_prompt="You are helpful")
"""

# Uncomment when ready to use OpenAI
"""
# Python standard library
import time
import logging
from typing import List, Optional

# Third-party imports
import openai  # pip install openai

# Local imports
from ..base import LLMProvider, LLMResponse, LLMError

# Configure logging
logger = logging.getLogger(__name__)


class OpenAIProvider(LLMProvider):
    '''
    OpenAI LLM provider implementation.
    
    Purpose:
        Provide standardized LLM interface for OpenAI GPT models.
        Supports GPT-4, GPT-3.5-turbo, and other OpenAI models.
    
    Features:
        - Automatic token counting
        - Streaming support (future)
        - Function calling (future)
        - Vision models support (future)
    
    Configuration:
        api_key: OpenAI API key (required)
        model_name: Model to use (default: gpt-3.5-turbo)
        organization: OpenAI organization ID (optional)
        timeout: Request timeout in seconds (default: 60)
    '''
    
    def __init__(
        self,
        model_name: str = "gpt-3.5-turbo",
        api_key: Optional[str] = None,
        organization: Optional[str] = None,
        timeout: int = 60,
        **kwargs
    ):
        '''
        Initialize OpenAI provider.
        
        Args:
            model_name: OpenAI model (gpt-4, gpt-3.5-turbo, etc.)
            api_key: OpenAI API key (or set OPENAI_API_KEY env var)
            organization: OpenAI organization ID (optional)
            timeout: Request timeout in seconds
            **kwargs: Additional configuration
        '''
        # Call parent constructor
        super().__init__(
            model_name=model_name,
            base_url=None,  # OpenAI uses default endpoint
            api_key=api_key,
            **kwargs
        )
        
        # Set API key
        if api_key:
            openai.api_key = api_key
        
        # Set organization if provided
        if organization:
            openai.organization = organization
        
        # Store timeout
        self.timeout = timeout
        
        # Log initialization
        logger.info(
            f"OpenAIProvider initialized: model={model_name}, "
            f"org={organization or 'default'}"
        )
    
    def chat(
        self,
        message: str,
        system_prompt: Optional[str] = None,
        temperature: float = 0.7,
        max_tokens: Optional[int] = None,
        **kwargs
    ) -> LLMResponse:
        '''
        Generate chat completion using OpenAI.
        
        Args:
            message: User message/prompt
            system_prompt: System prompt for context (optional)
            temperature: Response randomness (0.0-2.0)
            max_tokens: Maximum response tokens
            **kwargs: Additional OpenAI-specific parameters
        
        Returns:
            LLMResponse: Standardized response with text, tokens, timing
        
        Raises:
            LLMError: If generation fails
        '''
        # Validate parameters
        self._validate_parameters(
            temperature=temperature,
            max_tokens=max_tokens
        )
        
        # Build messages array
        messages = []
        if system_prompt:
            messages.append({"role": "system", "content": system_prompt})
        messages.append({"role": "user", "content": message})
        
        # Record start time
        start_time = time.time()
        
        try:
            # Call OpenAI API
            response = openai.ChatCompletion.create(
                model=self.model_name,
                messages=messages,
                temperature=temperature,
                max_tokens=max_tokens,
                timeout=self.timeout,
                **kwargs
            )
            
            # Calculate latency
            latency_ms = (time.time() - start_time) * 1000
            
            # Extract response text
            response_text = response.choices[0].message.content
            
            # Get token usage
            tokens_used = response.usage.total_tokens
            
            # Create standardized response
            return LLMResponse(
                text=response_text,
                model=self.model_name,
                tokens_used=tokens_used,
                latency_ms=latency_ms,
                metadata={
                    'provider': 'openai',
                    'finish_reason': response.choices[0].finish_reason,
                    'prompt_tokens': response.usage.prompt_tokens,
                    'completion_tokens': response.usage.completion_tokens
                },
                success=True,
                error=None
            )
        
        except Exception as e:
            # Calculate latency even on failure
            latency_ms = (time.time() - start_time) * 1000
            
            # Log error
            logger.error(
                f"OpenAI chat failed: {e}",
                exc_info=True,
                extra={
                    'model': self.model_name,
                    'message_length': len(message),
                    'latency_ms': latency_ms
                }
            )
            
            # Raise standardized error
            raise LLMError(
                message=f"OpenAI generation failed: {str(e)}",
                provider="openai",
                original_error=e
            )
    
    def health_check(self, force: bool = False) -> bool:
        '''
        Check if OpenAI API is accessible.
        
        Args:
            force: Force fresh check (ignore cache)
        
        Returns:
            bool: True if healthy, False otherwise
        '''
        try:
            # Try to list models as health check
            models = openai.Model.list(timeout=5)
            
            # Check if our model is available
            available_models = [m.id for m in models.data]
            is_healthy = self.model_name in available_models
            
            # Log health status
            if force:
                logger.info(
                    f"OpenAI health check: {'healthy' if is_healthy else 'unhealthy'} "
                    f"(model={self.model_name})"
                )
            
            return is_healthy
        
        except Exception as e:
            # Log error
            logger.error(f"OpenAI health check failed: {e}")
            return False
    
    def get_available_models(self) -> List[str]:
        '''
        Get list of available OpenAI models.
        
        Returns:
            List[str]: List of model names
        '''
        try:
            # Get models from OpenAI
            models = openai.Model.list(timeout=10)
            
            # Extract model IDs
            model_ids = [m.id for m in models.data]
            
            # Filter for chat models
            chat_models = [
                m for m in model_ids
                if 'gpt' in m.lower()
            ]
            
            # Log result
            logger.debug(f"Found {len(chat_models)} OpenAI chat models")
            
            return chat_models
        
        except Exception as e:
            # Log error
            logger.error(f"Failed to get OpenAI models: {e}")
            
            # Return common models as fallback
            return ['gpt-4', 'gpt-3.5-turbo']
"""

# Template marker - remove this when uncommenting above
print("OpenAI provider template - uncomment to use")
